Charity Water homage 1st draft. Still needs some
work. However, this update also takes advantage of the newer updates to jekyll that allows
assets to be placed inline with normal content and if they have certain markers they will
be compiled differently.

Compass had to be replaced with Bourbon for this to work because compass has it's own
compiling engine which does not have this nice feature.